+++ tests/testthat/test-read-oneshot.R
@@ -610,6 +610,29 @@ test_that("date-range", {
 
   expect_s3_class(returned_object$data, "tbl")
 })
+test_that("empty-dataset", {
+  testthat::skip_on_cran()
+  expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."
+
+  returned_object <-
+    redcap_read_oneshot(
+      redcap_uri            = credential$redcap_uri,
+      token                 = credential$token,
+      datetime_range_begin  = Sys.time(),
+      verbose               = FALSE
+    )
+
+  expect_equal(returned_object$data, expected=tibble::tibble(), label="The returned tibble should be empty", ignore_attr = TRUE) # dput(returned_object$data)
+  expect_equal(returned_object$status_code, expected=200L)
+  expect_equal(returned_object$raw_text, expected="", ignore_attr = TRUE) # dput(returned_object$raw_text)
+  expect_true(returned_object$records_collapsed=="", "A subset of records was not requested.")
+  expect_true(returned_object$fields_collapsed=="", "A subset of fields was not requested.")
+  expect_equal(returned_object$filter_logic, "")
+  expect_match(returned_object$outcome_message, regexp=expected_outcome_message, perl=TRUE)
+  expect_true(returned_object$success)
+
+  expect_s3_class(returned_object$data, "tbl")
+})
 test_that("guess_max-Inf", {
   testthat::skip_on_cran()
   expected_outcome_message <- "\\d+ records and \\d+ columns were read from REDCap in \\d+(\\.\\d+\\W|\\W)seconds\\."